requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
cattrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
dynaconf==3.1.9
exceptiongroup==1.0.0rc8
idna==3.3
iniconfig==1.1.1
packaging==21.3
pip==22.1.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest-dependency==0.5.1
pytest==7.1.2
pytz==2022.1
requests==2.28.1
setuptools==49.2.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.10
